+++ frontend/src/components/ScheduleTable.vue
