+++ frontend/src/components/Sidebar.vue
